[Refactor] Decouple the wallet class
